requirements.txt
Django>=4.1.3
djangorestframework>=3.13.1
django-cors-headers>=3.13.0
psycopg2>=2.9.3
Pillow>=9.2.0